requirements.txt
openai
streamlit
langfuse
python-dotenv
pycaret
matplotlib
seaborn
numpy